requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
propcache==0.4.1
proxmoxer==2.2.0
python-dotenv==1.1.1
python-telegram-bot==22.5
requests==2.32.5
sniffio==1.3.1
urllib3==2.5.0
yarl==1.22.0
